Fix hanging transport layer shutdown in tests
The transport layer is writing messages that it receives over the
network to a channel, from which the node implementation reads them.
When the node is shutting down, it stops reading incoming messages from
this channel. However, if there are still incoming messages on the wire
at that time, the transport layer tries to write them to the channel and
blocks (since nobody reads any more). This prevented a clean shutdown of
the transport layer.

Now the connecting node explicitly cancels the context of the connection
when disconnecting and the node being connected uses a select statement,
waiting for this event alongside trying to write the message to the
processing channel.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>