Implement PBFT View change and organize ISS files
The PBFT view change protocol using signed ViewChange messages is
implemented, including the re-transmission of Preprepare messages the
new leader might be missing.

The PBFT code is also better organized in files and logical blocks.
While the code itself is now organized much more logically and the
event handlers are sorted to represent the logical flow of the protocol,
many comments and explanations are still missing.

One function (deterministic list iteration) is added that uses generics
and thus go 1.18+ is necessary to compile the code. The GitHub Actions
are modified accordingly.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>