Implement DSL modules
The high-level goal is to make it possible to implement
distributed protocols in mir with a syntax that is very close to
the standard pseudocode notations as well as to separate the
protocol logic from the boilerplate. The way it is done is
basically by creating a domain-specific language inside go.

The motivating example can be found in pkg/bcb/bcbmodule.go.
The core implementation is in pkg/dsl/dslmodule.go.
The rest is mostly boilerplate and auxiliary functions.