Pass context by reference
The goal is to prevent the programmer from making a bug where they
pass the context by value when they send a request, but accept it
by reference in the handler (or vice versa). This would make the
handler not match the response event.
This modification does not completely eradicate the issue, but
makes it much harder to make such a mistake.