Add Batch and Transaction IDs to the mempool interface
The idea is to encapsulate batch and transaction IDs in the
mempool module to allow for more flexibility (e.g., the ID of a
batch can be a simple hash or a Merkle tree root) and avoid
potential subtle bugs (e.g., different modules could accidentally
use different hash functions). It also helps to make the code
more expressive.