Add multisig collector prototype
The prototype implements the desired interface and is supposed to
work properly in the crash-stop model. However, it does not use
a real persistent storage (and, hence, does not support node
recovery).
The prototype implements a Byzantine fault-tolerant protocol, but
it is susceptible to DOS attacks from Byzantine nodes.
Finally, at the moment, the prototype assumes reliable delivery of
messages between correct nodes.