Update all dependencies
Previously, the project wouldn't build with the latest release of
Go at the time (1.19). The reason for it is that the version of
the quic-go package was not compatible with go 1.19.
Unfortunately, older versions of other packages were not
compatible with the latest version of quic-go, so I just updated
all of them.