Only accept signature shares from distinct nodes in Recover
Upon looking at the tbls implementation, it is clear that it is not ready
to handle multiple/duplicate signature shares from the same node.

This is something that can be done ahead of time with greater efficiency
leading to less calls to `Recover`.

Note: anonymous threshold signature schemes cannot validate share origin,
and therefore may require modifying our requirements