Add support for logging into multiple files.
The interceptor is passed a new function as a creation parameter that
takes as input the interceptor's slice of events to be written on
file and returns a slice of slices of events (one per file). If
more than one, it closes the current file and creates a new file
to continue logging.