Clean up event log recorder and add Sqlite writer
The event log recorder can now be parametrized at instantiation
with one of two writers: Gzip and Sqlite.
The parameter is itself a function that returns the writer.

Some obsolete recorder options have been removed,
e.g. the compression level, as it is specific to the Gzip writer
and thus should be (and is) specified together with it.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>