Work around Interceptor nil comparison (Go quirk)
Avoid nil dereference if Intercept is called on a nil *Recorder
and simply do nothing. This can happen if a pointer type to *Recorder
is assigned to a variable with the interface type Interceptor.
Mir would treat that variable as non-nil,
thinking there is an interceptor, and call Intercept() on it.
For more explanation, see
https://mangatmodi.medium.com/go-check-nil-interface-the-right-way-d142776edef1

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>