Refactor checkpoint instantiation and epoch data
The epoch configuration is now passed directly to the checkpoint
protocol at creation, rather than in a separate event.
The relevant code is refactored to remove redundancy.

The epoch data is augmented by the membership of the epoch immediately
preceding the epoch to which the checkpoint is associated.
It is this membership that created the checkpoint
(the checkpoint of epoch e is created by nodes of epoch e-1)

The preceding membership needs to be used to verify checkpoint
certificates. This is not the case yet and must be addressed
in a future commit.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>