Make sure Interceptor works even if it is nil
Since the Interceptor is an interface type, it can happen that a nil
value of a concrete type is used in the node and, consequently,
Intercept() will be called on nil. The implementation of the concrete
type thus must make sure that calling Intercept even on the nil value
does not cause any problems.
For more explanation, see https://mangatmodi.medium.com/go-check-nil-interface-the-right-way-d142776edef1

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>