Fix subtle bug in state catchup protocol
Before, nodes from the current membership were monitored for the need
for state transmission, and the latest stable checkpoint was being sent
to them. This sometimes resulted in sending checkpoints to nodes that
were not yet in the transmitted checkpoint's membership, causing them
to get stuck.

This has been fixed by making sure we only send a checkpoint to a node
if the node is in that checkpoint's membership, and by the receiving
nodes ignoring checkpoints if they are not included in the corresponding
membership (such checkpoints can always be sent by faulty nodes).

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>