Work around protobuf/CBOR serialization issue
This is a DIRTY DIRTY hack. It works around the problem of protocol
buffers not distinguishing between an empty slice and a nil value
(both get serialized to the same bytes and deserialized to nil).
CBOR, however does make this distinction. The checkpoint data
structure does contain slices that may be empty
(DeliveredTXs inside ClientProgress). Thus, if a checkpoint were
directly serialized locally, and the very same checkpoint
transmitted to another node (which implies protobuf marshalling and
unmarshalling) and serialized there, the two outputs (local and
remote) would not be the same (an empty slice locally and nil on
the remote node). To keep the serialization consistent, we always
perform a protobuf marshalling-unmarshalling of the checkpoint.
Note that we cannot use the protobuf serialization directly either,
as it is not deterministic. This is very very inefficient and
something should be done about it.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>