Make mempool aware of the current epoch
This is required for not emitting the same transactions multiple times
in the same epoch, but still re-emitting transactions
that have not been delivered in a previous epoch.

The main changes to the logic of the mempool itself
are not yet included in this commit to simplify review.
They will follow in the next commit.

Signed-off-by: Matej Pavlovic <matopavlovic@gmail.com>